implemented parent dto for remote attributes, adapted exisiting dtos
